refactor(web): starts OSK preprocessor
fix(web): fixes kmwRecorder
