+++ web/source/osk/visualKeyboard.ts
@@ -1,5 +1,6 @@
 /// <reference path="activeLayout.ts" />
 /// <reference path="../utils/version.ts" />
+/// <reference path="preProcessor.ts" />
 
 namespace com.keyman.osk {
   let Codes = com.keyman.text.Codes;
@@ -911,7 +912,6 @@ namespace com.keyman.osk {
      *
      */
     touch: (e: TouchEvent) => void = function(this: VisualKeyboard, e: TouchEvent) {
-      let Processor = com.keyman.singleton.textProcessor;
       // Identify the key touched
       var t = <HTMLElement> e.changedTouches[0].target, key = this.keyTarget(t);
 
@@ -952,7 +952,7 @@ namespace com.keyman.osk {
       // Special function keys need immediate action
       if(keyName == 'K_LOPT' || keyName == 'K_ROPT')      {
         window.setTimeout(function(this: VisualKeyboard){
-          Processor.clickKey(key);
+          PreProcessor.clickKey(key);
         }.bind(this),0);
         this.keyPending = null;
         this.touchPending = null;
@@ -962,7 +962,7 @@ namespace com.keyman.osk {
         let touchProbabilities = this.getTouchProbabilities(e.changedTouches[0]);
         // While we could inline the execution of the delete key here, we lose the ability to
         // record the backspace key if we do so.
-        Processor.clickKey(key, e.changedTouches[0], this.layerId, touchProbabilities);
+        PreProcessor.clickKey(key, e.changedTouches[0], this.layerId, touchProbabilities);
         this.deleteKey = key;
         this.deleting = window.setTimeout(this.repeatDelete,500);
         this.keyPending = null;
@@ -971,7 +971,7 @@ namespace com.keyman.osk {
         if(this.keyPending) {
           this.highlightKey(this.keyPending, false);
           let touchProbabilities = this.getTouchProbabilities(this.touchPending);
-          Processor.clickKey(this.keyPending, this.touchPending, this.layerId, touchProbabilities);
+          PreProcessor.clickKey(this.keyPending, this.touchPending, this.layerId, touchProbabilities);
           this.clearPopup();
           // Decrement the number of unreleased touch points to prevent
           // sending the keystroke again when the key is actually released
@@ -1047,7 +1047,7 @@ namespace com.keyman.osk {
         // Output character unless moved off key
         if(this.keyPending.className.indexOf('hidden') < 0 && tc > 0 && !beyondEdge) {
           let touchProbabilities = this.getTouchProbabilities(e.changedTouches[0]);
-          Processor.clickKey(this.keyPending, e.changedTouches[0], this.layerId, touchProbabilities);
+          PreProcessor.clickKey(this.keyPending, e.changedTouches[0], this.layerId, touchProbabilities);
         }
         this.clearPopup();
         this.keyPending = null;
@@ -1321,10 +1321,8 @@ namespace com.keyman.osk {
      *  Repeat backspace as long as the backspace key is held down
      **/
     repeatDelete: () => void = function(this: VisualKeyboard) {
-      let Processor = com.keyman.singleton.textProcessor;
-
       if(this.deleting) {
-        Processor.clickKey(this.deleteKey);
+        PreProcessor.clickKey(this.deleteKey);
         this.deleting = window.setTimeout(this.repeatDelete,100);
       }
     }.bind(this);
@@ -1763,7 +1761,7 @@ namespace com.keyman.osk {
       // Process as click if mouse button released anywhere over key
       if(util.eventType(e) == 'mouseup') {
         if(key.id == this.currentKey) {
-          keyman.textProcessor.clickKey(key);
+          PreProcessor.clickKey(key);
         }
         this.currentKey='';
       }
